agentica/requirements.txt
pydantic
fastapi
uvicorn
python-dotenv

mcp
langchain
langchain-community
langgraph
langchain-google-genai
langchain-openai
langchain-anthropic
langgraph-checkpoint-sqlite
langchain-chroma
xai-sdk
duckduckgo-search<7.0.0
chromadb
toolbox-core

structlog
pytest
pytest-asyncio
pytest-cov
rich
